requirements.txt
flask
flask-cors
python-dotenv
langchain
google-generativeai
